Rework auto-save behaviour to avoid Neotree glitch
There is a rather annoying UI glitch when focusing Neotree after
modifying a file. The BufLeave event which triggers the auto-save
behaviour sometimes causes Neotree to refresh itself and miss cursor
motions already inputted. This causes the cursor to "jump back" a node
or two on the tree.

This is especially noticeable when spamming a motion in quick succession
after focusing the tree.

To address this issue the auto-save behaviour has been reworked such
that only modified, _hidden_ buffers are saved during the BufLeave and
now BufEnter events. This means that if Neotree is focused while the
currently being worked on buffer is still visible, it will not be saved.
If a _new_ buffer is then opened (replacing the modified one) then it
will now be hidden and therefore auto-saved. As this happens on
BufEnter, Neotree is no longer focused and no longer susceptible to
glitching.

The FocusLost event still results in all modified buffers (visible or
not) to be written.