Improve inter-process yank behaviour
Currently when yanking in one neovim instance and pasting in another the
paste always needs to be cycled one back as the process registers still
container whatever was in them before the yank.

This same property is true when starting a neovim instance after yanking
in another one - though I can't for the life of me figure out why.

This sets up an autocmd to populate the process registers from the
history item at the top of the stack when focus is gained.

Additionally the registers are populated from the _second_ history item
on boot. Why it needs to be the second I have no idea.