Use single-file mode in clojure-lsp
When clojure-lsp is started in a project without a detectable project
root then start the clojure-lsp without a root-dir configured
(effectively enabling single-file mode).

This is done over the previous approach of starting clojure-lsp in a tmp
directory to ensure that syntax highlighting still works for files
'outside the project root'.