Allow language server to control buffer validation
This fixes an issue with jdtls where go-to-definition for jdt:// uri's
would fail to attach as the opened buffer would not pass the default
buffer validation.

This adds an optional buf_is_valid fn on the server config which allows
handing off validation to the specific server.

For jdtls this function just returns true as jdtls can, so far, handle
all java files I've encountered.