Replace neo-tree with nvim-tree.lua
I have found neo-tree to slow down neovim to the point that neovim needs
to be restarted to be able to do anything.

This happens over time while more and more files get opened in neotree.

This replaces my neotree integration with nvim-tree which seems to not
suffer from this problem.