Replace metadata-based sql injection with regex
This uses regex to match if a string contains SQL which is then in turn
used to apply the sql language injection to the string.

This works much better than the previous ^:sql metadata approach as this
doesn't require adding anything to source-code and works for ANY string
in source that might contain sql.