Rework buffer-switcher to be shared across windows
Currently the buffer switcher maintains a separate stack of buffers per
window.

I actually prefer being able to switch to the previous buffer regardless
of which window it happened to be in.

This makes the stack global, but includes the window in each entry and
on switch, also focuses the relevant window.